+++ lib/guardian/plug/verify_header.ex
@@ -36,6 +36,9 @@ if Code.ensure_loaded?(Plug) do
       `:none` will not use a prefix.
     * `key` - The location to store the information in the connection. Defaults to: `default`
     * `halt` - Whether to halt the connection in case of error. Defaults to `true`.
+    * `secret` - The secret used to verify the token. Accepts any value supported by
+      the token module, or a one-arity function that receives the connection and
+      returns the secret.
     * `:refresh_from_cookie` - Looks for and validates a token found in the request cookies. (default `false`)
 
     Refresh from cookie option
@@ -89,7 +92,8 @@ if Code.ensure_loaded?(Plug) do
            module <- Pipeline.fetch_module!(conn, opts),
            claims_to_check <- Keyword.get(opts, :claims, %{}),
            key <- storage_key(conn, opts),
-           {:ok, claims} <- Guardian.decode_and_verify(module, token, claims_to_check, opts) do
+           verify_opts <- Guardian.Plug.put_secret_from_conn(opts, conn),
+           {:ok, claims} <- Guardian.decode_and_verify(module, token, claims_to_check, verify_opts) do
         conn
         |> Guardian.Plug.put_current_token(token, key: key)
         |> Guardian.Plug.put_current_claims(claims, key: key)